Separa plugin principale dalla GUI in file distinti
- istat_boundaries_downloader.py: mantiene solo IstatBoundariesDownloader
  con entry point, configurazione e avvio del dialogo
- istat_boundaries_downloader_dialog.py: nuovo file con DownloaderDialog
  contenente tutta la logica UI e di download
